SU-4510: Fixing windows install issue
